chapter_02/Exercise_2.07_stateful_applications/container/requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
distlib==0.3.7
filelock==3.12.3
platformdirs==3.10.0
psycopg2-binary==2.9.10
virtualenv==20.24.4